requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
requests
pandas
scikit-learn
asyncpg
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
httpx
